--- assembly/src/parsers/blocks.rs
@@ -1,76 +1,0 @@
-use super::{CodeBlock, Operation, Vec};
-use vm_core::{utils::group_vector_elements, DecoratorList};
-
-// UTILITY FUNCTIONS
-// ================================================================================================
-
-pub fn combine_blocks(mut blocks: Vec<CodeBlock>) -> CodeBlock {
-    // merge consecutive Span blocks.
-    let mut merged_blocks: Vec<CodeBlock> = Vec::with_capacity(blocks.len());
-    // Keep track of all the consecutive Span blocks and are merged together when
-    // there is a discontinuity.
-    let mut contiguous_spans: Vec<CodeBlock> = Vec::new();
-
-    blocks.drain(0..).for_each(|block| {
-        if block.is_span() {
-            contiguous_spans.push(block);
-        } else {
-            if !contiguous_spans.is_empty() {
-                merged_blocks.push(combine_spans(&mut contiguous_spans));
-            }
-            merged_blocks.push(block);
-        }
-    });
-    if !contiguous_spans.is_empty() {
-        merged_blocks.push(combine_spans(&mut contiguous_spans));
-    }
-
-    // build a binary tree of blocks joining them using Join blocks
-    let mut blocks = merged_blocks;
-    while blocks.len() > 1 {
-        let last_block = if blocks.len() % 2 == 0 {
-            None
-        } else {
-            blocks.pop()
-        };
-
-        let mut grouped_blocks = Vec::new();
-        core::mem::swap(&mut blocks, &mut grouped_blocks);
-        let mut grouped_blocks = group_vector_elements::<CodeBlock, 2>(grouped_blocks);
-        grouped_blocks.drain(0..).for_each(|pair| {
-            blocks.push(CodeBlock::new_join(pair));
-        });
-
-        if let Some(block) = last_block {
-            blocks.push(block);
-        }
-    }
-
-    blocks.remove(0)
-}
-
-/// Returns a CodeBlock [Span] from sequence of Span blocks provided as input.
-pub fn combine_spans(spans: &mut Vec<CodeBlock>) -> CodeBlock {
-    if spans.len() == 1 {
-        return spans.remove(0);
-    }
-
-    let mut ops = Vec::<Operation>::new();
-    let mut decorators = DecoratorList::new();
-    spans.drain(0..).for_each(|block| {
-        if let CodeBlock::Span(span) = block {
-            for decorator in span.decorators() {
-                decorators.push((decorator.0 + ops.len(), decorator.1.clone()));
-            }
-            for batch in span.op_batches() {
-                ops.extend_from_slice(batch.ops());
-            }
-        } else {
-            panic!(
-                "Codeblock was expected to be a Span Block, got {:?}.",
-                block
-            );
-        }
-    });
-    CodeBlock::new_span_with_decorators(ops, decorators)
-}
