+++ assembly/src/parsers/tests.rs
@@ -1,9 +1,8 @@
-use super::{parse_module, parse_program, BTreeMap, Instruction, LocalProcMap, Node, ProcedureAst};
-use crate::{
-    parsers::ast::{ModuleAst, ProgramAst},
-    ProcedureId,
+use super::{
+    parse_module, parse_program, BTreeMap, Felt, Instruction, LocalProcMap, ModuleAst, Node,
+    ProcedureAst, ProcedureId, ProgramAst,
 };
-use vm_core::{Felt, FieldElement};
+use crate::{ONE, ZERO};
 
 // UNIT TESTS
 // ================================================================================================
@@ -12,7 +11,7 @@ use vm_core::{Felt, FieldElement};
 #[test]
 fn test_ast_parsing_program_simple() {
     let source = "begin push.0 assertz end";
-    let values: Vec<Felt> = vec![Felt::ZERO];
+    let values: Vec<Felt> = vec![ZERO];
     let nodes: Vec<Node> = vec![
         Node::Instruction(Instruction::PushConstants(values)),
         Node::Instruction(Instruction::Assertz),
@@ -198,19 +197,19 @@ fn test_ast_parsing_module_nested_if() {
 
     let mut procedures: LocalProcMap = BTreeMap::new();
     let proc_body: Vec<Node> = vec![
-        Node::Instruction(Instruction::PushConstants([Felt::ONE].to_vec())),
+        Node::Instruction(Instruction::PushConstants([ONE].to_vec())),
         Node::IfElse(
             [
-                Node::Instruction(Instruction::PushConstants([Felt::ZERO].to_vec())),
-                Node::Instruction(Instruction::PushConstants([Felt::ONE].to_vec())),
+                Node::Instruction(Instruction::PushConstants([ZERO].to_vec())),
+                Node::Instruction(Instruction::PushConstants([ONE].to_vec())),
                 Node::IfElse(
                     [
-                        Node::Instruction(Instruction::PushConstants([Felt::ZERO].to_vec())),
+                        Node::Instruction(Instruction::PushConstants([ZERO].to_vec())),
                         Node::Instruction(Instruction::Sub),
                     ]
                     .to_vec(),
                     [
-                        Node::Instruction(Instruction::PushConstants([Felt::ONE].to_vec())),
+                        Node::Instruction(Instruction::PushConstants([ONE].to_vec())),
                         Node::Instruction(Instruction::Sub),
                     ]
                     .to_vec(),
@@ -267,23 +266,23 @@ fn test_ast_parsing_module_sequential_if() {
 
     let mut procedures: LocalProcMap = BTreeMap::new();
     let proc_body: Vec<Node> = vec![
-        Node::Instruction(Instruction::PushConstants([Felt::ONE].to_vec())),
+        Node::Instruction(Instruction::PushConstants([ONE].to_vec())),
         Node::IfElse(
             [
                 Node::Instruction(Instruction::PushConstants([Felt::new(5)].to_vec())),
-                Node::Instruction(Instruction::PushConstants([Felt::ONE].to_vec())),
+                Node::Instruction(Instruction::PushConstants([ONE].to_vec())),
             ]
             .to_vec(),
             vec![],
         ),
         Node::IfElse(
             [
-                Node::Instruction(Instruction::PushConstants([Felt::ZERO].to_vec())),
+                Node::Instruction(Instruction::PushConstants([ZERO].to_vec())),
                 Node::Instruction(Instruction::Sub),
             ]
             .to_vec(),
             [
-                Node::Instruction(Instruction::PushConstants([Felt::ONE].to_vec())),
+                Node::Instruction(Instruction::PushConstants([ONE].to_vec())),
                 Node::Instruction(Instruction::Sub),
             ]
             .to_vec(),
@@ -380,7 +379,7 @@ of the comments is correctly parsed. There was a bug here earlier."
 
     let proc_body_baz: Vec<Node> = vec![
         Node::Instruction(Instruction::PadW),
-        Node::Instruction(Instruction::PushConstants(vec![Felt::ZERO])),
+        Node::Instruction(Instruction::PushConstants(vec![ZERO])),
     ];
     let docs_baz =
         "Test documenation for export procedure baz in parsing test. Lorem ipsum dolor sit amet,
